ЛИСТОК ИДИ КА ТЫ НА ХУЙ

+++ src/main/java/dimas/model/User.java
@@ -1,12 +1,16 @@
 package dimas.model;
+
 import javax.persistence.*;
+import javax.validation.constraints.Min;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.Size;
 
 @Entity
 @Table(name = "users")
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
     @Column(name = "name")
     private String name;
     @Column(name = "age")
